# api/fomo-swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "description": "Endpoints labeled open do not require authentication. The protected endpoints on the other hand, do require\nauthentication. I'm not sure how long the jwt token should last. I'm thinking we should set the expire on\nthat token to be super short - like 5 minutes (upto an hour maybe?) to minimize the amount of time an\nattacker can use that token. The refresh token will last longer - currently 7 days. If you make a request\nto a protected endpoint using a \"Refresh\" token in your request headers, you will receive a new\nauthorization token (set-authorization) and refresh token (set-refresh) in the response headers when\nyou make a request with an expired authorization token. You MUST replace both tokens in your request headers\nto stay authenticated. The old refresh token gets replaced on the backend therefore, you need to use the\nnew refresh token to remain actively logged in.",
    "title": "FOMO API",
    "version": "0.0.1"
  },
  "host": "stage.fomo.exchange",
  "basePath": "/api",
  "paths": {
...
    "/balances": {
      "get": {
        "description": "Returns all balances for the user. Use optional query param 'symbol' as lowercase ticker symbol - e.g. ada.\nNote: This is going away once accounts is fully implemented.",
        "tags": [
          "balances"
        ],
        "summary": "get all balances (protected)",
        "operationId": "getAllBalanceClient",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "Symbol",
            "name": "symbol",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "\"data\" will contain array of balances with \"status\": \"success\"",
            "schema": {
              "$ref": "#/definitions/ResponseBalancesSuccess"
            }
          },
          "500": {
            "description": "the message will state what the internal server error was with \"status\": \"error\"",
            "schema": {
              "$ref": "#/definitions/responseError"
            }
          }
        }
      }
    },
    "/devices": {
      "get": {
        "description": "Returns a list of the user's registered devicesr.",
        "tags": [
          "devices"
        ],
        "summary": "all registered devices (protected)",
        "operationId": "GetAllDevices",
        "responses": {
          "200": {
            "description": "\"data\" will contain array of devices with \"status\": \"success\"",
            "schema": {
              "$ref": "#/definitions/ResponseDevicesSuccess"
            }
          },
          "500": {
            "description": "the message will state what the internal server error was with \"status\": \"error\"",
            "schema": {
              "$ref": "#/definitions/responseError"
            }
          }
        }
      },
      "post": {
        "description": "Registers a new device for a user so they may receive push notifications.",
        "tags": [
          "devices"
        ],
        "summary": "add new device (protected)",
        "operationId": "AddDevice",
        "parameters": [
          {
            "x-go-name": "DeviceType",
            "description": "Required.",
            "name": "deviceType",
            "in": "body",
            "schema": {
              "type": "string"
            }
          },
          {
            "x-go-name": "DeviceToken",
            "description": "Required.",
            "name": "deviceToken",
            "in": "body",
            "schema": {
              "type": "string"
            }
          },
          {
            "x-go-name": "ExternalDeviceID",
            "description": "Required.",
            "name": "externalDeviceID",
            "in": "body",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "\"data\" will be non null with \"status\": success",
            "schema": {
              "$ref": "#/definitions/ResponseDeviceSuccess"
            }
          },
          "400": {
            "description": "missing params",
            "schema": {
              "$ref": "#/definitions/responseError"
            }
          },
          "500": {
            "description": "the message will state what the internal server error was with \"status\": error",
            "schema": {
              "$ref": "#/definitions/responseError"
            }
          }
        }
      }
    },
    "/devices/:deviceID": {
      "get": {
        "description": "Get a user's device by the device's ID.",
        "tags": [
          "devices"
        ],
        "summary": "get a device by ID (protected)",
        "operationId": "GetDevice",
        "responses": {
          "200": {
            "description": "\"data\" will contain device stuffs with \"status\": \"success\"",
            "schema": {
              "$ref": "#/definitions/ResponseDeviceSuccess"
            }
          },
          "500": {
            "description": "the message will state what the internal server error was with \"status\": constRes.Error",
            "schema": {
              "$ref": "#/definitions/responseError"
            }
          }
        }
      },
      "put": {
        "description": "Updates a user's device.",
        "tags": [
          "devices"
        ],
        "summary": "update a registered device (protected)",
        "operationId": "UpdateDevice",
        "parameters": [
          {
            "x-go-name": "DeviceType",
            "description": "Required.",
            "name": "deviceType",
            "in": "body",
            "schema": {
              "type": "string"
            }
          },
          {
            "x-go-name": "DeviceToken",
            "description": "Required.",
            "name": "deviceToken",
            "in": "body",
            "schema": {
              "type": "string"
            }
          },
          {
            "x-go-name": "ExternalDeviceID",
            "description": "Required.",
            "name": "externalDeviceID",
            "in": "body",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "\"data\" will contain updated device info with \"status\": success",
            "schema": {
              "$ref": "#/definitions/ResponseDeviceSuccess"
            }
          },
          "500": {
            "description": "the message will state what the internal server error was with \"status\": error",
            "schema": {
              "$ref": "#/definitions/responseError"
            }
          }
        }
      },
      "delete": {
        "description": "Removes device by ID.",
        "tags": [
          "devices"
        ],
        "summary": "removes a user's device (protected)",
        "operationId": "DeleteDevice",
        "responses": {
          "200": {
            "description": "data will be null with \"status\": success",
            "schema": {
              "$ref": "#/definitions/responseSuccess"
            }
          },
          "500": {
            "description": "the message will state what the internal server error was with \"status\": error",
            "schema": {
              "$ref": "#/definitions/responseError"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "AccountBalances": {
      "type": "object",
      "properties": {
        "balances": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Balance"
          },
          "x-go-name": "Balances"
        }
      },
      "x-go-package": "github.com/asciiu/gomo/api/controllers"
    },
...
    "ApiDevice": {
      "type": "object",
      "properties": {
        "deviceID": {
          "type": "string",
          "x-go-name": "DeviceID"
        },
        "deviceToken": {
          "type": "string",
          "x-go-name": "DeviceToken"
        },
        "deviceType": {
          "type": "string",
          "x-go-name": "DeviceType"
        },
        "externalDeviceID": {
          "type": "string",
          "x-go-name": "ExternalDeviceID"
        }
      },
      "x-go-package": "github.com/asciiu/gomo/api/controllers"
    },
...
    "Balance": {
      "type": "object",
      "properties": {
        "available": {
          "type": "number",
          "format": "double",
          "x-go-name": "Available"
        },
        "balanceID": {
          "type": "string",
          "x-go-name": "BalanceID"
        },
        "currencyName": {
          "type": "string",
          "x-go-name": "CurrencyName"
        },
        "exchange": {
          "type": "string",
          "x-go-name": "ExchangeName"
        },
        "exchangeAvailable": {
          "type": "number",
          "format": "double",
          "x-go-name": "ExchangeAvailable"
        },
        "exchangeLocked": {
          "type": "number",
          "format": "double",
          "x-go-name": "ExchangeLocked"
        },
        "exchangeTotal": {
          "type": "number",
          "format": "double",
          "x-go-name": "ExchangeTotal"
        },
        "keyID": {
          "type": "string",
          "x-go-name": "KeyID"
        },
        "locked": {
          "type": "number",
          "format": "double",
          "x-go-name": "Locked"
        }
      },
      "x-go-package": "github.com/asciiu/gomo/api/controllers"
    },
...
    "Device": {
      "type": "object",
      "properties": {
        "deviceID": {
          "type": "string",
          "x-go-name": "DeviceID"
        },
        "deviceToken": {
          "type": "string",
          "x-go-name": "DeviceToken"
        },
        "deviceType": {
          "type": "string",
          "x-go-name": "DeviceType"
        },
        "externalDeviceID": {
          "type": "string",
          "x-go-name": "ExternalDeviceID"
        }
      },
      "x-go-package": "github.com/asciiu/gomo/api/controllers"
    },
...
    "DeviceList": {
      "type": "object",
      "properties": {
        "devices": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ApiDevice"
          },
          "x-go-name": "Devices"
        }
      },
      "x-go-package": "github.com/asciiu/gomo/api/controllers"
    },
...
    "ResponseBalancesSuccess": {
      "type": "object",
      "title": "A ResponseBalancesSuccess will always contain a status of \"successful\".",
      "properties": {
        "data": {
          "$ref": "#/definitions/AccountBalances"
        },
        "status": {
          "type": "string",
          "x-go-name": "Status"
        }
      },
      "x-go-package": "github.com/asciiu/gomo/api/controllers"
    },
    "ResponseDeviceSuccess": {
      "type": "object",
      "title": "A ResponseDeviceSuccess will always contain a status of \"successful\".",
      "properties": {
        "data": {
          "$ref": "#/definitions/UserDeviceData"
        },
        "status": {
          "type": "string",
          "x-go-name": "Status"
        }
      },
      "x-go-package": "github.com/asciiu/gomo/api/controllers"
    },
    "ResponseDevicesSuccess": {
      "type": "object",
      "title": "A ResponseDevicesSuccess will always contain a status of \"successful\".",
      "properties": {
        "data": {
          "$ref": "#/definitions/DeviceList"
        },
        "status": {
          "type": "string",
          "x-go-name": "Status"
        }
      },
      "x-go-package": "github.com/asciiu/gomo/api/controllers"
    },
...
    "UserData": {
      "type": "object",
      "properties": {
        "devices": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Device"
          },
          "x-go-name": "Devices"
        },
        "user": {
          "$ref": "#/definitions/UserInfo"
        }
      },
      "x-go-package": "github.com/asciiu/gomo/api/controllers"
    },
    "UserDeviceData": {
      "type": "object",
      "properties": {
        "device": {
          "$ref": "#/definitions/ApiDevice"
        }
      },
      "x-go-package": "github.com/asciiu/gomo/api/controllers"
    },
    "UserInfo": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string",
          "x-go-name": "Email"
        },
        "first": {
          "type": "string",
          "x-go-name": "First"
        },
        "last": {
          "type": "string",
          "x-go-name": "Last"
        },
        "userID": {
          "type": "string",
          "x-go-name": "UserID"
        }
      },
      "x-go-package": "github.com/asciiu/gomo/api/vendor/github.com/asciiu/gomo/user-service/models"
    },
...
    "responseError": {
      "description": "This response may or may not include data encapsulating the user information.",
      "type": "object",
      "title": "A ResponseSuccess will always contain a status of \"successful\".",
      "properties": {
        "message": {
          "type": "string",
          "x-go-name": "Message"
        },
        "status": {
          "type": "string",
          "x-go-name": "Status"
        }
      },
      "x-go-name": "ResponseError",
      "x-go-package": "github.com/asciiu/gomo/api/controllers"
    },
...
    "responseSuccess": {
      "type": "object",
      "title": "A ResponseSuccess will always contain a status of \"successful\".",
      "properties": {
        "data": {
          "$ref": "#/definitions/UserData"
        },
        "status": {
          "type": "string",
          "x-go-name": "Status"
        }
      },
      "x-go-name": "ResponseSuccess",
      "x-go-package": "github.com/asciiu/gomo/api/controllers"
    }
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}